meta-mxt-*: app-usbcam-http: hack support for non-USB cameras
The issue is that the first camera found, will be used, regardless if it
works or not.

For CSI2 cameras, we will need a script to do the setup of the sensor at
boot. And these cameras don't support the VIDIOC_S_PARM ioctl(), so we
ignore the failure (and log it) if it happens.

Signed-off-by: Alexandru Ardelean <alex@shruggie.ro>